+++ bld/dip/javavm/c/jvmsupp.c
@@ -252,7 +252,7 @@ dip_status      GetAddrCue( struct mad_jvm_findaddrcue_acc *acc, struct mad_jvm_
     return( DCAssignLocation( &dst, &src, sizeof( *ret ) ) );
 }
 
-unsigned NameCopy( char *buff, const char *src, unsigned buff_size, unsigned len )
+size_t NameCopy( char *buff, const char *src, size_t buff_size, size_t len )
 {
     if( buff_size > 0 ) {
         --buff_size;
