+++ bld/dip/javavm/c/jvmtype.c
@@ -291,11 +291,11 @@ int DIGENTRY DIPImpTypeCmp( imp_image_handle *ii, imp_type_handle *it1,
     return( 0 );
 }
 
-unsigned DIGENTRY DIPImpTypeName( imp_image_handle *ii, imp_type_handle *it,
-                unsigned num, symbol_type *tag, char *buff, unsigned buff_size )
+size_t DIGENTRY DIPImpTypeName( imp_image_handle *ii, imp_type_handle *it,
+                unsigned num, symbol_type *tag, char *buff, size_t buff_size )
 {
     char        *p;
-    unsigned    len;
+    size_t      len;
 
     *tag = ST_NONE;
     if( num != 0 ) return( 0 );
