+++ bld/dip/skel/c/skelmod.c
@@ -51,8 +51,8 @@ walk_result     DIGENTRY DIPImpWalkModList( imp_image_handle *ii,
    return( WR_CONTINUE );
 }
 
-unsigned        DIGENTRY DIPImpModName( imp_image_handle *ii,
-                        imp_mod_handle im, char *buff, unsigned buff_size )
+size_t DIGENTRY DIPImpModName( imp_image_handle *ii,
+                        imp_mod_handle im, char *buff, size_t buff_size )
 {
     //TODO:
     /*
